Refactor marker position state
